requirements.txt
transformers
torch
librosa
matplotlib
dtw-python
numpy